Make just a function, not a Mixin
