+++ terra/tests/test_compute_utils.py
@@ -1,3 +1,4 @@
+import os
 from unittest import mock
 import warnings
 
@@ -42,7 +43,9 @@ class TestComputeUtilsCase(TestCase):
                                         clear=True))
     super().setUp()
     # Register this Compute
-    settings.configure({'compute': {'arch': Compute.__module__}})
+    settings.configure({'compute': {'arch': Compute.__module__},
+                        'processing_dir': self.temp_dir.name})
+        # 'test_dir': '/opt/projects/terra/terra_dsm/external/terra/foo'})
 
     # Manually register without using decorator
     Compute.register(Service)(Service_test)
@@ -118,3 +121,78 @@ class TestComputeHandler(TestComputeUtilsCase):
 
   # Make sure this can be run twice
   test_compute_handler2 = test_compute_handler
+
+
+# Dummy mock to double check args
+def mock_popen(*args, **kwargs):
+  return (args, kwargs)
+
+
+class TestBaseJust(TestComputeUtilsCase):
+  def setUp(self):
+    self.patches.append(mock.patch.object(utils, 'Popen', mock_popen))
+    super().setUp()
+
+    # Make a copy
+    self.original_env = os.environ.copy()
+
+  def tearDown(self):
+    super().tearDown()
+    # Make sure nothing inadvertently changed environ
+    self.assertEqual(self.original_env, os.environ)
+
+  def test_just_simple(self):
+    default_justfile = os.path.join(os.environ['TERRA_TERRA_DIR'], 'Justfile')
+    # Call just, and get the args calculated, retrieved via mock
+    args, kwargs = utils.just("foo  bar")
+    self.assertEqual(args, (('bash', 'just', 'foo  bar'),))
+    self.assertEqual(set(kwargs.keys()), {'executable', 'env'})
+    self.assertEqual(kwargs['env']['JUSTFILE'], default_justfile)
+
+  def test_just_custom_env(self):
+    default_justfile = os.path.join(os.environ['TERRA_TERRA_DIR'], 'Justfile')
+    # Use the env kwarg
+    args, kwargs = utils.just("foo", "bar", env={"FOO": "BAR"})
+    self.assertEqual(args, (('bash', 'just', 'foo', 'bar'),))
+    self.assertEqual(set(kwargs.keys()), {'executable', 'env'})
+    self.assertTrue(kwargs.pop('executable').endswith('bash'))
+    self.assertEqual(kwargs, {'env': {'FOO': 'BAR',
+                                      'JUSTFILE': default_justfile}})
+
+  def test_just_custom_justfile(self):
+    # Use the justfile kwarg
+    args, kwargs = utils.just("foobar", justfile="/foo/bar")
+    self.assertEqual(args, (('bash', 'just', 'foobar'),))
+    self.assertEqual(set(kwargs.keys()), {'executable', 'env'})
+    self.assertEqual(kwargs['env']['JUSTFILE'], "/foo/bar")
+
+  def test_just_kwargs(self):
+    default_justfile = os.path.join(os.environ['TERRA_TERRA_DIR'], 'Justfile')
+    # Use the shell kwarg for Popen
+    args, kwargs = utils.just("foobar", shell=False)
+    self.assertEqual(args, (('bash', 'just', 'foobar'),))
+    self.assertEqual(set(kwargs.keys()), {'executable', 'env', 'shell'})
+    self.assertEqual(kwargs['shell'], False)
+    self.assertEqual(kwargs['env']['JUSTFILE'], default_justfile)
+
+  def test_logging_code(self):
+    # Test the debug1 diffdict log output
+    with self.assertLogs(utils.__name__, level="DEBUG1") as cm:
+      env = os.environ.copy()
+      env.pop('PATH')
+      env['FOO'] = 'BAR'
+        # Sometimes JUSTFILE is set, so make this part of the test!
+      with mock.patch.dict(os.environ, JUSTFILE='/foo/bar'):
+        utils.just("foo", "bar", env=env)
+
+    env_lines = [x for x in cm.output if "Environment Modification:" in x][0]
+    env_lines = env_lines.split('\n')
+    self.assertEqual(len(env_lines), 5, env_lines)
+
+    # Verify logs say PATH was removed
+    self.assertTrue(any(o.startswith('- PATH:') for o in env_lines))
+    # FOO was added
+    self.assertTrue(any(o.startswith('+ FOO:') for o in env_lines))
+    # JUSTFILE was changed
+    self.assertTrue(any(o.startswith('+ JUSTFILE:') for o in env_lines))
+    self.assertTrue(any(o.startswith('- JUSTFILE:') for o in env_lines))
